feat: Move ds-form to 3-unit and update docs
