+++ website/src/content/docs/4-page/grid.md
@@ -7,6 +7,13 @@ description: Learn about grid components in DS one
 
 Grid components create structured layouts using CSS Grid and provide alignment utilities for organizing content.
 
+## Flexible Content
+
+`ds-grid` is a **flexible layout component** that can hold any content. Each grid cell can contain:
+- **`ds-card`** - For card-based layouts
+- **`ds-container`** - For nested containers
+- Any other components or HTML elements
+
 ## Features
 
 The grid components include:
@@ -18,3 +25,13 @@ The grid components include:
 - Alignment and distribution options
 
 Grids are essential for creating responsive, multi-column layouts with precise control.
+
+## Example
+
+```html
+<ds-grid columns="3" gap="medium">
+  <ds-card>Card 1</ds-card>
+  <ds-card>Card 2</ds-card>
+  <ds-card>Card 3</ds-card>
+</ds-grid>
+```
